Add getPackImageUrl helper to MusicLibrary
